feat: api key 스키마 파일 이동
