+++ apps/client/src/entities/home/index.ts
@@ -1,1 +1,2 @@
 export * from './model/types';
+export * from './model/schema';
