+++ apps/client/src/entities/home/model/types.ts
@@ -1,5 +1,4 @@
 import { ApiResponse } from '@repo/shared/types';
-import z from 'zod';
 
 export interface ApiKeyData {
   apiKey: string;
@@ -7,18 +6,6 @@ export interface ApiKeyData {
   scopes: string[];
   description: string;
 }
-
-export const ApiKeySchema = z.object({
-  scopes: z.array(z.string()).min(1, { message: '권한 범위를 최소 1개 이상 선택해주세요.' }),
-  description: z.string().min(1, { message: '설명을 입력해주세요.' }),
-});
-
-export type ApiKeyType = z.infer<typeof ApiKeySchema>;
-
-export type CreateApiKeyType = z.infer<typeof ApiKeySchema>;
-
-export type UpdateApiKeyType = z.infer<typeof ApiKeySchema>;
-
 export interface AvailableScopeData {
   scope: string;
   description: string;
